Scramble plugins now almost completely working. In order to implement multi-slice for cubes, a scramble attributes feature was added. Scrambles can now have an unlimited number of booleans attributed with them, and CCT will automatically create the checkboxes necessary for each attribute. The status of all these attributes is saved upon exiting.
All that remains is to fix server scrambles, which have been broken.

Also the gui's minimum size has been seriously reduced. The reasoning behind this was to prevent the program resizing itself while being used, which is a very disconcerting behavior to the end-user. Instead, CCT will simply truncate any information that doesn't fit, and it will be up to the user to resize the window should they wish to.