Removed a System.out.println() from SolveTime.java
Statistics.java. I ran eclipse's code formatter on this after I had made a few of changes. As a result, the diff is going to look very crazy. Sorry 'bout that. We should probably agree on a coding style and just automatically force it to conform to that standard in the future.