MASSIVE scramble rehaul (overhaul, whatever). Everything is now a ScramblePlugin, ScrambleVariation, or ScrambleCustomization. I discourage any other naming, as it will only lead to confusion.
There is still some minor cleanup I noticed could be done, although it will be easy to do. One example: I'm not sure that ScrambleViewComponents need to keep track of all the colorschemes, I think that should be changed.

The configuration related stuff inside of the three main classes of this commit (ScramblePlugin, Variation, and Customization) may get relocated.

The ScrambleCustomization gui stuff is not implemented (in configdialog) yet, although it will be soon.

I also need to finish the scrambledebugger, which should be really easy now that I've written ScramblePlugin.