Forgot to mention in previous commit: -comment area now has some informational text -added comment columns to the times table and sessions table.
Noticed that I was adding the instance of cct as a table change listener too early, which would cause the speaker to speak all the times loaded from the database.

Also, selecting the most recent time and scrolling to the bottom are apparently broken, and I can't figure out why.