Scrambles should now be fully secured. The scramblelist now holds strings, instead of scrambles, which allowed me to remove the imported boolean from scramble.java, and the setAttributes() method.
I am worried that a malicious scramble plugin could access and modify the final arrays in another plugin to kill it. I need to analyze about this, and see if it's a potential security exploit. Probably would be best to just prohibit inter-scramble plugin interaction. Not sure how easy that will be though...

The undo-redo count seems to be broken. I'll try to fix that next.